requirements.txt
aliyun-python-sdk-core==2.13.13
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.9.0
amqp==2.6.0
appnope==0.1.2
backcall==0.1.0
billiard==3.5.0.5
black==22.6.0
bump2version==1.0.1
bumpversion==0.6.0
celery==4.0.0
certifi==2019.9.11
chardet==3.0.4
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-aliyun-oss2-storage==0.1.5
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-filter==2.2.0
django-formtools==2.1
django-import-export==2.0
django-reversion==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
environs==6.1.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.15.0
idna==2.8
importlib-metadata==1.7.0
ipython==7.12.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.10.3
jmespath==0.9.4
kombu==4.6.11
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
marshmallow==3.2.2
mypy-extensions==0.4.3
odfpy==1.4.0
openpyxl==3.0.2
oss2==2.9.1
parso==0.6.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycryptodome==3.9.4
Pygments==2.5.2
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
tomli==1.2.3
traitlets==4.3.3
typed-ast==1.5.4
typing-extensions==4.1.1
uritemplate==3.0.1
urllib3==1.25.7
uWSGI==2.0.20
vine==1.3.0
wcwidth==0.1.8
xlrd==1.2.
XlsxWriter==1.2.6
xlwt==1.3.0
zipp==3.1.0
